+++ product-service/src/main/java/br/edu/atitus/product_service/clients/CurrencyClient.java
@@ -1,13 +1,15 @@
 package br.edu.atitus.product_service.clients;
 
+import java.math.BigDecimal;
+
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 
-@FeignClient(name = "currency-service", fallback = CurrencyFallback.class) // url = "http://localhost:8100" (antiga url)
+@FeignClient(name = "currency-service", fallback = CurrencyFallback.class)
 public interface CurrencyClient {
 
 	@GetMapping("/currency/{value}/{source}/{target}")
-	CurrencyResponse getCurrency(@PathVariable double value, @PathVariable String source, @PathVariable String target);
-
+	CurrencyResponse getCurrency(@PathVariable BigDecimal value, @PathVariable String source,
+			@PathVariable String target);
 }
